Fixes logs when making a tyypo in a command
